+++ CHANGELOG.md
@@ -10,6 +10,13 @@
 
 ### Removed
 
+## [1.1.0] - 2020-02-15
+
+### Changed
+
+- set history as global state
+- accuracy becomes the max pointer size by type
+
 ## [1.0.1] - 2020-02-14
 
 ### Added